+++ src/Game/GameObjects/BrickWall.cpp
@@ -86,7 +86,7 @@ void BrickWall::render()const
 	renderBrick(EBlockLocation::BottomRight);
 }
 
-void BrickWall::update(const uint64_t delta)
+void BrickWall::update(const double delta)
 {
 
 }
