+++ src/Game/GameObjects/Water.h
@@ -25,7 +25,7 @@ public:
 
 	Water(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
 	virtual void render()const override;
-	virtual void update(const uint64_t delta)override;
+	virtual void update(const double delta)override;
 
 private:
 	void renderBlock(const EBlockLocation eBlockLocation) const;
